requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
Pillow
drf-yasg
django-filter
